Analyzing data on the Star Wars movies.
While waiting for Star Wars: The Force Awakens, the team at
FiveThirtyEight was interested in answering some questions about Star
Wars fans. One question that particularly interested the team was: Does
the rest of America realize that “The Empire Strikes Back” is clearly
the best of the bunch?

The team needed to collect data before they could get started answering
this question. They used SurveyMonkey, an online survey tool, to survey
Star Wars fans. They received 835 responses total, which you can find
here.

In this project, we'll be cleaning and exploring the dataset in Jupyter.